requirements.txt
Django
django-admin
django-excel-response2
django-js-asset
django-mptt
django-six
django-widget-tweaks
excel-base
freetype-py
nest-asyncio
packaging
parso
pdfminer
pdfminer.six
pickleshare
Pillow
platformdirs
prompt-toolkit
psutil
pure-eval
pycairo
pycparser
pycryptodome
Pygments
PyJWT
PyPDF2
python-dateutil
pyzmq
reportlab
requests
rlPyCairo
soupsieve
sqlparse
stack-data
TimeConvert
tornado
traitlets
typing_extensions
tzdata
tzlocal
urllib3
wcwidth
xlwt
